+++ assembly/src/assembler/mod.rs
@@ -572,12 +572,12 @@ impl Assembler {
         let wrapper_proc = self.module_graph.get_procedure_unsafe(gid);
         let proc = wrapper_proc.unwrap_ast().unwrap_procedure();
         let proc_body_id = if num_locals > 0 {
-            // for procedures with locals, we need to update fmp register before and after the
-            // procedure body is executed. specifically:
+            // For procedures with locals, we need to update fmp register before and after the
+            // procedure body is executed. Specifically:
             // - to allocate procedure locals we need to increment fmp by 4 times the number of
-            //   locals
-            // - to deallocate procedure locals we need to decrement it by the same amount We leave
-            // 4 elements between locals to properly support reading and writing words to locals.
+            //   locals, and
+            // - to deallocate procedure locals we need to decrement it by the same amount. We leave
+            //   4 elements between locals to properly support reading and writing words to locals.
             let locals_frame = Felt::from(num_locals * WORD_SIZE as u16);
             let wrapper = BodyWrapper {
                 prologue: vec![Operation::Push(locals_frame), Operation::FmpUpdate],