+++ miden/tests/integration/operations/io_ops/mem_ops.rs
@@ -117,7 +117,7 @@ fn mem_stream() {
     let inputs = [1, 2, 3, 4, 5, 6, 7, 8];
 
     // the state is built by replacing the values on the top of the stack with the values in memory
-    // addresses 0 and 4 (i.e., 1 through 8). Thus, the first 8 elements on the stack will be 1
+    // addresses `[0..8)`. Thus, the first 8 elements on the stack will be 1
     // through 8 (in stack order, with 8 at stack[0]), and the remaining 4 are untouched (i.e., 9,
     // 10, 11, 12).
     let state: [Felt; 12] =